requirements.txt
azure-identity==1.6.0
azure-mgmt-subscription==3.1.1
azure-mgmt-resource==20.0.0
azure-mgmt-web==7.1.0